src/PythonPttServer/pyproject.toml
[project]
name = "pythonpttserver"
version = "0.1.0"
description = "gRPC PTT server for Zilean in Python"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "rich (>=14.0.0,<15.0.0)",
    "rank-torrent-name (>=1.9.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

